Implemented Testing on Firefox Browser
