+++ coverage-frontend/summary/index.html
@@ -23,30 +23,30 @@
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">59.83% </span>
+                <span class="strong">58.64% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>146/244</span>
+                <span class='fraction'>139/237</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
                 <span class="strong">50% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>34/68</span>
+                <span class='fraction'>32/64</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">51.42% </span>
+                <span class="strong">48.48% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>18/35</span>
+                <span class='fraction'>16/33</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">59.74% </span>
+                <span class="strong">58.54% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>141/236</span>
+                <span class='fraction'>137/234</span>
             </div>
         
             
@@ -80,17 +80,17 @@
 </thead>
 <tbody><tr>
 	<td class="file high" data-value="bookings.js"><a href="bookings.js.html">bookings.js</a></td>
-	<td data-value="96.22" class="pic high">
-	<div class="chart"><div class="cover-fill" style="width: 96%"></div><div class="cover-empty" style="width: 4%"></div></div>
+	<td data-value="95.95" class="pic high">
+	<div class="chart"><div class="cover-fill" style="width: 95%"></div><div class="cover-empty" style="width: 5%"></div></div>
 	</td>
-	<td data-value="96.22" class="pct high">96.22%</td>
-	<td data-value="106" class="abs high">102/106</td>
-	<td data-value="84.21" class="pct high">84.21%</td>
-	<td data-value="38" class="abs high">32/38</td>
+	<td data-value="95.95" class="pct high">95.95%</td>
+	<td data-value="99" class="abs high">95/99</td>
+	<td data-value="88.23" class="pct high">88.23%</td>
+	<td data-value="34" class="abs high">30/34</td>
 	<td data-value="100" class="pct high">100%</td>
-	<td data-value="11" class="abs high">11/11</td>
-	<td data-value="98" class="pct high">98%</td>
-	<td data-value="100" class="abs high">98/100</td>
+	<td data-value="9" class="abs high">9/9</td>
+	<td data-value="95.91" class="pct high">95.91%</td>
+	<td data-value="98" class="abs high">94/98</td>
 	</tr>
 
 <tr>
@@ -146,7 +146,7 @@
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2024-01-28T09:22:04.406Z
+                at 2024-01-29T01:40:32.731Z
             </div>
         <script src="prettify.js"></script>
         <script>
