+++ coverage-frontend/summary/login.js.html
@@ -247,7 +247,7 @@ function viewUser() {
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2024-01-28T09:22:04.406Z
+                at 2024-01-29T01:40:32.731Z
             </div>
         <script src="prettify.js"></script>
         <script>
